Working build again.
Background: Build broke because cargo by default seems to ignore the Cargo.lock file (what's the point of having it then?) and one of the dependencies had been updated and required a newer version of rust. Adding --locked seemed like it would make sense to make cargo actually use the lock-file, but that doesn't seem to work with the two step build process of first building the dependencies, then building the application (you end up with a "Hello, world!" program in that case). I've now upgraded the rust base image to a newer version which fixes the build. But I don't like the fact that the build may break in the future without any code changes.